backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "passlib[argon2]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyhamcrest>=2.1.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
]
